+++ src/main/java/com/github/k1mb1/vkr_backend/group/web/response/SubgroupResponse.java
@@ -7,7 +7,7 @@ import lombok.Builder;
 @Builder
 public record SubgroupResponse(
     UUID id,
-    Short index,
+    Integer index,
     Instant createdAt,
     Instant updatedAt
 ) {}
